Moved helper functions into separate files and updated gocstrings
